requirements.txt
dateutil
django>=3.0
django_json_widget>=1.1.1
django-admin-list-filter-dropdown>=1.0.2
django-import-export
python-telegram-bot>=13.14